Version 0.1 is deprecated.
The current system uses an arduino UNO for real-time sensor control, and the Raspberry Pi 3 for database management and web interface display. The RPi.GPIO module in Python is not suitable for real time control. Under the limited use of the original system it performed adequately, but as the intervals on the sensor queries became shorter, a different solution was needed and the Raspberry Pi could no longer be used by itself.